requirements.txt
Django==3.2.12
django-cors-headers==3.8.0
django-csp==3.7
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2
